requirements.txt
python-dotenv==1.0.1
black==24.4.0
isort==5.13.2
langchain==0.1.17
langchain-core
langchain-openai==0.1.3
langgraph==0.0.38
grandalf==0.8
langchain-community
